requirements.txt
python-dotenv
streamlit
huggingface_hub
TTS
pydub
replicate
openai